Add option to disable favicon filtering in nginx
By default nginx server template filters out requests for favicon.ico,
so that they do not appear in the access logs. Becauce of that, if
favicon.ico is served to the user by different means (for example,
application server itself), request is not passed through. Adding
'favicon: False' to server configuration options will disable this
functionality and allow serving of favicon.ico as normal.