Fix idempotency issue within 'apt' role
On a normal ginas host with only common roles enabled this bug did not
show up, but it might be present later, when different services are
enabled.

As it turns out (actually, I haven't bothered to check, just a guess),
'apt_repository' module, to ensure that a specified repository is not
present in /etc/apt/sources.list, rewrites the whole file. I'm not sure
why, but a line with one '#' character (creating a better whitespace in
the comment text) is modified by the module during rewrite process to
'# ', ie. with an added space after '#' character.

This innocent rewrite presents a problem on the next run - Ansible
'template' module sees that /etc/apt/sources.list file has been modifed,
and replaces it with a new, correct version (without the space present
in '#' empty line). Which generates a change event. Which then shows up
at the end of the playbook run with usually just 1 change.

And then user might wonder, what changed? The playbook is big, it might
take several minutes to find out what caused the change. And it also
means, that the whole playbook is idempotent, which shouldn't happen.

I pondered for a while if that should be reported as a bug in
'apt_repository' module, after all this is the culprit, not 'template'
module which only does its job. But in the end it's not that important,
because it's just a meaningless empty comment line that is modified.

So I decided to remove that lonely '#' character instead and move on.

/ducks/
                        ___
                      .'   '.__
                     (    o {__)     __  '-.  .'
                      \  ._.'    __.'  '.   '   .-'
               /\      )  (     (__} o   ) .-')    _
              |  ``'"'`    \       '.__. /` .'   .'  '
              \  {`'-       }    (`--')  _.'--`\
            jgs'. '-._     /      `-.{        _ )
           ``^^`^`^`^^`^^`^``^`^````^`^^^`^`^^`^^`^^