Replace 'ansible_managed' with a static comment
'ansible_managed' variable in templates was causing Ansible to rewrite
the template despite the fact that configuration haven't changed. This
rewrite triggered a "changed" status, which notified some handlers,
which were needlessly restarted. All that caused repeaded playbook runs
to be not idempotent, longer and could distrupt services for some time.

This patch replaces 'ansible_managed' variable in all templates with
a short static comment which indicates that the file is managed by
Ansible and all changes made to it will be lost.