Add data file for the project backup script
This file provides list of repositories to back up by the backup script.

To allow easy updates of the list, backup script will download current
version of the backup-data.txt file with current list of repositories.
To prevent tampering, this file is GnuPG-signed and backup script will
check the validity of the signature before proceeding with the backup.