Change how 'debops.ruby' role manages Ruby version
From now on, Ruby will fall back on APT to select the version of Ruby
packages to install - if Ruby 2.1 is available, APT should automatically
install it; if not, APT will install Ruby packages already present in
the distribution.

Ruby backports can be enabled by setting ruby_version: 'backport'.
This functionality might be moved in the future to a separate playbook.