Add an 'archive' capability in Postfix
If 'archive' capability is enabled on a Postfix server, it configures
transport map and {sender,recipient}_bcc_maps to automatically add blind
carbon copy address to each message passing through the SMTP server.
This in effect creates a copy of the message which is sent to an archive
server (which needs to be configured separately).

Address of the archive account and server to send messages to is by
default configured automatically. If 'local' mail is enabled, archived
messages will be directed to 'mail-archive' account on the SMTP host. If
'local' mail is disabled, archived mail messages will be sent to
'mail-archive' account to a server on local domain with 'archive.'
subdomain. Both of these (archive server with a DNS subdomain and mail
account) need to be configured separately by the sysadmin.

You can also change how 'archive' capability behaves using several
variables configured via inventory. You can change the destination mail
address, specify how archived mail is sorted (if at all) and only
archive mail from or to specific domains, among other things.