Add a way to disable accept rule via variable
You can add 'item.disabled: True/False' or 'item.enabled: True/False'
keys to 'dport_accept' iptables rules to be able to enable/disable rules
using Ansible variables.

For example, if you want to open the firewall if a certain list has
a list of IP addresses or networks, you can add a key to your ferm
variables similar to:

  item.enabled: '{{ ipaddr_list }}'

This will automatically enable or disable ferm rules if a specified
variable can be resolved to True or False. You can use 'item.disabled'
key to invert the result - firewall rules will be enabled if the
variable is resolved to False, and vice versa.