Convert role from common to group-based
Previously BoxBackup role was designed to be a common role ran from
"common.yml" playbook on all hosts. It has been now converted to
a group-based role which is designed to run on specific group of hosts,
"debops_boxbackup". First host in this group will act as a BoxBackup
server, the rest will be configured as clients. Default server can be
also changed in 'boxbackup_server' variable.