Automatically select one nginx server as default
Currently nginx server listens both on IPv4 and IPv6 networks, however
only IPv6 network is enabled by default. To make things simplier,
a directive 'ipv6only=off' which only needs to be present once, is tied
to 'default_server' directive.

Result of that decision is a requirement that at least one nginx server
must be set as the default, otherwise IPv4 traffic won't be accepted by
nginx. Code in this commit enables that function - it checks if a server
with name defined in 'nginx_default_name' variable is present among the
list of servers, if not, it checks if a server with 'item.default: True'
is present, otherwise the first server in 'nginx_servers' list is
selected as the default. The server name that has been picked is saved
in local facts, so that nginx role configures all servers correctly.