Add support for autoconfigured 6to4 tunnels
6to4 (https://en.wikipedia.org/wiki/6to4) is a method to tunnel IPv6
traffic in IPv4 packets. 'debops.ifupdown' role will check if default or
specified IPv4 address of a host is a public one - if it is and IPv6
network connectivity is not present, an 6to4 tunnel will be
automatically configured. These tunnels are configured only on public
IPv4 network, hosts behind NAT will not have 6to4 enabled and should
instead be configured via normal IPv6 router advertisements.