Add support for IP indexing in ipaddr() filter
If you specify a positive or a negative integer in an ipaddr() query, it
will be interpreted as an index and filter will return an IP address at
the specified index. Positive numbers count from the start of the
address range, negative numbers start from the end of the address range.
If index is out of range, False will be returned instead.