Use 'root:adm' permissions in ferm configuration
It seems that in Debian ferm configuration is set by default to root:adm
and with SGID bit set all config files in '/etc/ferm/' are created with
these permissions. After an upgrade config files are chowned again to
root:adm. To stop this back-and-forth, ferm role will use 'root:adm'
owner and group by default for all configuration files.