Don't create stub firewall for disabled domains
The idea before this change was that in case an user disabled a domain
(IPv4 or IPv6) using 'ferm_filter_domains', ferm role would still manage
a basic iptables/ip6tables firewall to protect the services on disabled
side. But since user can just disable for example IPv6 support in the
kernel, managing "stub" firewall is not necessary, or even can be
harmful when IPv6 is not enabled and ip6tables emits an error and stops
the whole Ansible run.