'debops.pki' rewrite
This is a redesign of the 'debops.pki' role, which manages SSL/TLS
certificates. This rewrite allows you to manage a complete CA chain,
from Root Certificate Authority, through Intermediate CA to
a specialized CA which signs certificates for other hosts.

This commit is a first iteration of the rewrite, more commits will
follow with updates to the code and documentation.