Change the way certificates are identified
Using IP addresses to bind certificates to host is not a good idea,
because it can leak information and IP addresses change. Instead, use
a hashed MAC address of primary network interface as an unique host
identifier. This lets us easily match a particular certificate to
a particular host using a simple table of hashed MAC addresses.