Better control over available git-shell-commands
Each user entry can now have a 'item.commands' key. It's a list of
commands allowed to be used by a given account, and default list can be
manipulated by either adding additional commands in a separate list, or
creating a difference (examples provided in comments).

'debops.gitusers' role will check what files are present in each
account ~/git-shell-commands/ directory and remove everything that is
not included in list of allowed commands.