Rewrite of debops.gitusers role
This is a rewrite of the 'debops.gitusers' role focused on making the
role more general and allowing for better flexibility in the future.

Shell scripts have been changed from a templated files stored directly
in user home directories to a set of normal files stored in central
location (/var/lib/gitusers/ by default). They are not templated and use
the ~/.giconfig file to read current user configuration and act
accordingly.

Users can now create different types of repositories - at the moment
a default repository type is 'jekyll' and it will finction similarly to
how GitHub Pages operates. Another available repository type is 'deploy'
and it allows to run user-provided scripts from 'deploy/' directory
included in the repository (if user has a correct permission for it).