Rewrite SASL Auth / Cyrus support
This rewrite changes the way SASL is implemented in 'debops.postfix' and
prepares the role for easy switch to Dovecot in the future.

Instead of "hijacking" the 'saslauthd' service, 'debops.postfix' will
configure a separate service for Postfix authentication using support
included in Debian. More information can be found on
https://wiki.debian.org/PostfixAndSASL