Vastly simplify conditions by removing `is defined` checks.
The variable `ifupdown_static_config` was already always defined,
except if `ifupdown` is false and thus check_static.yml was skipped.
So checking for it's existence is superfluous in all `ifupdown and ...`
conditions (due to short-circuit evaluation). Only in `not ifupdown
and ...` conditions, the existence of the variable would have to be
tested.

For symmetry, we no define always define the variable, avoiding
the test or default in the later cases.