Disable nginx and ferm dependencies in debops.apt
debops.ferm role, which debops.postfix depends on, has some weird
interaction with debops.nginx role which is used in debops.apt role.
This interfaction results in ferm restart handler not firing properly
when notified by Postfix role, only skipped. Which resulted in this fix:

https://github.com/debops/ansible-postfix/commit/0ef60f524ad045c3197caee1b40721be7740e35c

Now to reverse that, I decided to disable nginx and ferm in debops.apt
role. It is used by Debian Preseed setup, which should be moved to
a separate role by now anyway, so this will force me to create that role
soon. ferm is used by 'apt-cacher-ng' which also should be moved to
a separate role.

If either of those are needed, the dependencies can be temporarily
enabled to configure needed services and then disabled again.