Fix error when admin in LXC containers is disabled
When 'lxc_template_admin_name' was disabled by setting it as False,
Ansible would emit an error about joining string and bool values. So
instead, make sure that a string value is always present in
'lxc_template_admin_home'. It shouldn't ever be used, but just in case,
make it a legit one, as well.