Reorder the tasks so secret variables are present
Role uses a shared username / password combination on all managed hosts,
but because at the moment Ansible fails when creating the same file
multiple times, they need to be created only once. Therefore a separate
'set_fact' task is used.

This change reorders the task so that variables are populated earlier
and are available for '/etc/snmp/snmpd.conf' template.