Bypass issues with vanishing files during backups
Add 'rsnapshot_cmd_rsync' variable which allows you to set the path to
a script executed during backup.

Add 'rsync-no-vanished' wrapper script, which bypasses an issue when
`files during long backups vanish`_ and the backup cycle is broken. It's
set to be run by default for all hosts, can be overriden per host.