Add support for SSH key lookup in external sources
Support for 'AuthorizedKeysCommand' option will be disabled by default,
and can be enabled on Debian Jessie as well as on Debian Wheezy with
backported OpenSSH version using 'sshd_authorized_keys_lookup' variable.
Scripts that perform the lookups will be executed on a separate system
UNIX account to provide privilege separation.