Revert `option` markup. Fixed to pass `assert_docs` in test framework.
About the error:

```
Warning, treated as error:
/home/travis/build/debops/ansible-atd/docs/getting-started.rst:62: WARNING: u'option' reference target not
found: --tags
```

The `option` role tries to reference a previously given command line help, if one is available. But as `sphinx-build` is run with `-n Run in nit-picky mode, warn about all missing references.` during testing it becomes a problem.

http://sphinx-doc.org/markup/inline.html#role-option

I disabled the `option` stuff which is preferred over disabling the `-n`
option.