Rename 'ferm' variable to 'ferm_enabled'.
This change is needed to avoid issues with Ansible templating the 'ferm'
package in lists with contents of the 'ferm' variable.

If you have 'ferm' disabled anywhere (set to 'False'), you will need to
change the name of the variable in inventory to the new one before running
this role. Otherwise there should be no changes necessary.