Add a way to prevent changes in live interfaces
By setting 'ifupdown_reconfigure_auto' variable to 'False' you can
prevent the role from messing with the live network configuration, but
still configure the interfaces in '/etc/network/interfaces'. This is
useful on a production server with virtual machines or containers
running, since it prevents modification to network bridges which
requires restart of the network interfaces and may drop the existing
bridge layout.