Migrated role to DebOps and changed license to GPLv3.
The switch to GPLv3 was made to avoid uncertainty in regards if using
roles in playbooks is considered linking or not and thus if the
playbooks would also need to be under AGPLv3. Also, this role depends on
`debops.secrets` which is GPLv3.

But the better reason for the switch is to allow seamless copy/paste of
code. I know that it is nice to have DebOps as a set of best practice
roles to peek and reuse code. Although the aspect of the AGPL is important
to me, the copy paste thing also is.