Support realm/authority management by other roles
The 'pki_dependent_realms' and 'pki_dependent_authorities' lists can be
used by other Ansible roles to create PKI realms and authorities in the
playbook automatically.

The new 'item.when' parameter of PKI realm can be used to conditionally
enable or disable the realm on a specific host or group.