Better control over CA certificates
A specific internal CA can be omitted in the system CA certificates by
setting the 'item.system_ca' parameter to 'False'. The CA certificate
won't be trusted system-wide, but still will be available in the PKI
realm.

If you specify 'item.alt_authority' parameter in a given CA as a name of
another internal CA, the certificate of that internal CA will be copied
to the PKI realm as well, where it will be used instead as the one that
signed the certificate. This can be used to use different CA for signing
the server certificates, and different one for signing the client
certificates.