Change URL for fetching the postint.sh script.
The former url `<item>.seed` requires the user to add `*.seed` or
`<item>.seed` to her DNS - which she might not be able to do. The new
url was already accepted be the web server and does not require
changes to the DNS as it works with only `seed`. If one can not able
to add `seed` to the DNS, she might use "preseed_base_domain" to make
the server listen on a hostname available via DNS.