Added ``apt_preferences__preset_list`` for advanced users.
The presets are intended for advanced users who want to install software
from different releases but want to keep there main system on the stable
release.

@drybjed and me discoursed a bit how to add such examples to the role.
The previous consent was to add it to the documentation but I think
adding it to the defaults has some benefits in this case which are that
users can directly use the presets and benefit from updates directly.
Also, this avoids creating redundancy.

The preferences are tested by me for some time now.

Related to
https://github.com/ypid/ansible-packages/commit/cf59393274cf402a6e767c9eb8fc19888b0ba838.