Checked and fixed possible quoting problems using shellcheck.
There are still a number of open warnings reported by shellcheck but
they are not as easy to fix as quoting.

This PR is not really tested and I am not 100 % sure that I did not
introduce bugs by this rework. Unfortunately I did not find a tool to do
this quite monoton task for me.

Related to
https://github.com/debops/ansible-pki/pull/57#issuecomment-217175678