Explicitly set `!requiretty` for the `bootstrap__sudo_group`.
This ensures that `sudo` with `rsync` is allowed
for the `bootstrap__sudo_group` even when `requiretty`` has been
configured to be the default for users.

I use [a role to configure sudo](https://github.com/ypid/ansible-sudo) on my
systems and set `!requiretty` as default.

Form the manpage:

> If set, sudo will only run when the user is logged in to a real tty.  When this
> flag is set, sudo can only be run from a login session and not via other means
> such as cron(8) or cgi-bin scripts.  This flag is off by default.

So this patch is only required when the admin has changed the default from
`!requiretty` to `requiretty` and wants to use `sudo` with `rsync` as a user in
the `bootstrap__sudo_group`.

Related to: https://github.com/ansible/ansible/issues/4676#issuecomment-27605422