Ensure that highly sensitive files are not checked into version control.
@drybjed Might there be any other directories except `private` where
such files (not symlinks to them ;) ) could end up in?

I decided to add the ignore file pattern to this role directly to ensure
that everybody who might use version control for `/etc` gets this
default behaviour and the ignore file patterns are based on the
directory structure of the role.