Configure 'cron' entry in /etc/cron.d/
DebOps roles are deisnged as a sysadmin tool, and are oriented in such
a way as to not interfere with user-modifable space unless explicitly
requested (it might not be 100% true yet, though). This change should
ensure that the 'cron' entry managed by the role is maintained by 'root'
account and not the user account, so that user cannot modify it by
accident.