Fix nms set a default_server if default_server is top level domain.
nginx default template check if default_server with a test if default
server name is in the item nginx server name.
If item server name is a string instead of an array and if default
server name is TLD of nms server name they match (substring) by mistake.
This breaks nginx reload later on as now we have two default servers.