Remove index.php from try_files.
try_file is executed in the current context, "location /" in our
case, and would serve the php script as a static file.

Use $uri.php as the last argument in order to do an internal redirect
to the php location, which in turn will take care of returning a 404,
if that script doesn't exist.