Fixes default servername matches with subdomain.
if server name is a string, the in operator split its fqdn
and match succeed vs default_server if this string is a subdomain
of the proper default_server.

That is "default_server in item.name".
Fix by way of switching operands to get "item.name in default_server"
and account for item.name scalar or vector type.