Allow interfaces with /32 or /128 netmask
This is useful for special interfaces like dummy interfaces which add
an addtional IP address to the host which is not on any subnet attached
to this host. Eg. if you have floating service IPs.

The current code filters all IP address by the ipaddr("host/prefix")
filter. This filter discards values with a netmask of /32 or /128. This
change removes that filter to allow these addresses. The
ipaddr("address") filter which is applied later in the template still
ensures that no invalid addresses can be specified.

Fixes #54