Renamed postscreen template for main and master
The problem was that the main.cf's configuration template was included
in the generated master.cf.  That in turn led to invalid syntax.

My naive guess is that ansible is confused as to which template to take
because both main.cf/ and master.cf/ contain a file called
"postscreen.cf".  When renaming both, it seems to work better.

There may be a different solution like prefix each include with the
relevant subdirectory, i.e. "master.cf/" and "main.cf/" to prevent
further confusions, but I haven't tried that.