only active deny users config if 'deny' in dovecot_user_accounts
The task for generating the dovecot config currently checks for
'{{ "auth-deny.conf.ext" if "deny" in dovecot_user_accounts else [] }}'.
The template, however, does not perform the same check. Hence the task
does not update "auth-deny.conf.ext".  The consequence is that the
default file will be used which references /etc/dovecot/deny-users
instead of dovecot.deny.  The former, however, does not exist which
makes dovecot complain.

This change applies the same deny checking logic used in the task to the
template.  The auth-deny config should thus only be included if we
indeed manage deny users.