Added apache__redirect_to_https variable
Currently, each vhost can define "redirect_to_https".  If True, which is
the current default, then the Apache configuration will include a
Redirect line.

This is good, except that VirtualHost rules seem to have precedence over
other rules.  That is annoying when you want to define a global rule to
serve, say, ACME requests, because Apache would respect the VirtualHost
rules and redirect the plain HTTP request instead of serving the token.
To help debops users to not having to touch each and every defined
vhost, the new variable defines the default value for redirect_to_https,
which used to be "True".  The default is still "True", but now you can
override that easily.