Don’t use a offset by default for plain mode. This is a bad default!
Plain mode support has not been released yet. If you used this
unreleased version and setup non-swap devices with it, you will need to
set the offset manually so that anything which is on plaintext device
mapper target will get recognized correctly.
If you used used plain mode for swap, you don’t need to do anything. The
role will update your hosts on the next run.

It was believed that 2048 would be a good default value to support UUIDs
for the ciphertext block device used for swap. This does not work on
Debian however:

https://unix.stackexchange.com/questions/271097/setting-up-encrypted-swap-in-lmde-debian

(It could be made to work by setting `precheck=/bin/true` but please
don’t do this!)