Fix issue with parameter sorting algorithm
Previously new option values added to the dictionary had their weight
relative to the currently processed entry id. Due to that they eneded up at
the end of the value list because entries added through inventory have
much higher id than the default entries.

Now, each option value weight will be set relative to the existing id of
the given option if it's already present. This should ensure that the
weight of the values from inventory is close to the weight of values
from defaults, which makes sorting using custom weight do as expected.