Add note to Ansigenome template that separate repos are not upstream
I fear that people might miss the recent merge into the monorepo and
create PRs against the old repos which will cause unnecessary work.

Not sure if that will help but it is a start. I would propose to also
add a temporary CONTRIBUTING.rst file into the root of each separate
repo.