Add an improved script to check PEP8 compliance
The 'pycodestyle' command alone finds only '*.py' files. The new script
is a wrapper which finds all Python scripts in the repository, including
Ansible fact scripts, and checks their PEP8 compliance.

The script is not enabled on Travis or in the Makefile yet, because
there are 690 PEP8 errors still to fix in the repository.