Disable Transparent HugePages for Redis service
When the 'debops.redis' role is used to install the Redis service, it
will disable the support for Transparent HugePages via the
'debops.sysfs' role. THP might result in latency problems in Redis and
suggested fix is to disable that functionality on production hosts.

More details:
- https://redis.io/topics/latency
- https://stackoverflow.com/questions/42591511/

Fixes https://github.com/debops/ansible-redis/pull/22