Fix missing plugin paths in generated ansible.cfg
The 'debops' script will now include paths to Ansible plugins located in
the playbook directories, in addition to a set of separate directories.
This is done to preserve backwards compatibility with existing set of
DebOps playbooks.

The script will now look for Ansible roles, playbooks and plugins in the
'debops/' subdirectory of a project directory. This allows installation
of the DebOps monorepo inside of the project directory to keep it
separate from the main monorepo in '~/.local/share/debops/', either
directly or via a symlink, allowing for a completely self-contained
development environment.

The 'debops-tools' version is increased to allow clean upgrade on
installs done via PyPI.