Fix MS Windows support in Vagrantfile
The MS Windows operating system doesn't support symlinks, which are used
in the DebOps repository. Because of that, repository on MS Windows
hosts cannot be directly synced under the '/vagrant' directory.

Because of that, on MS Windows hosts, directory syncing will be
disabled, and DebOps monorepo will be installed via the 'debops-update'
command.