[debops.netbox] Fix 'become' and shell tasks
The 'become' keyword is required by Ansible on all tasks that also use
the 'decome_user' keyword.

The 'shell' and 'command' modules were recently changed to use the
default shell of the account that the task is executed with, for example
via 'become'. The 'netbox' account uses the '/usr/sbin/nologin' shell by
default, therefore a working shell needs to be specified explicitly.