[debops.librenms] Install MIBs conditionally
The 'snmp-mibs-downloader' APT package is located in the 'non-free'
section of the Debian Archive. The 'debops.librenms' role will install
it only if the 'non-free' repositories are enabled by the 'debops.apt'
role.

The LibreNMS application configured by the 'debops.librenms' role does
not currently work without the non-free MIBs installed. I'm not sure if
that can be disabled yet.