[debops.lxc] Fix idempotency issues with firewall
The 'debops.lxc' role uses the 'debops.ferm' role to configure the
firewall, specifically to enable packet forwarding to the LXC
containers. The 'debops.ferm' role uses currently two default variables
to track the state of the forwarding rules. Because the 'debops.lxc'
playbook used only one of these variables, the idempotency was broken on
the second run, and only worked afterwards.

This patch adds usage of the missing 'debops.ferm' variable in the
'lxc.yml' playbook, this should ensure idempotency after the first
playbook execution.