[debops.libvirtd] Replace ferm hook with a script
Current 'ferm' post-hook which reloads 'libvirtd' to load its firewall
rules breaks the boot sequence on hosts with 'systemd' - 'ferm.service'
tries to reload 'libvirtd.service' which isn't started at a given boot
stage and waits indefinitely.

This patch replaces the post-hook command with a script which checks, if
'libvirtd.service' is started and only then performs a reload.