[debops.gitlab_runner] Patch 'vagrant-libvirt'
When 'vagrant-libvirt' plugin is used to create a custom Vagrant Box
file and package it, it uses 'virt-sysprep' command to clear any
unnecessary files in the VM. This includes SSH host keys.

Unfortunately, at the moment, after the new VMs generated using this
custom Vagrant Box are started, they don't regenerate the SSH keys
automatically. This breaks Vagrant provisioning and the 'vagrant ssh'
commands.

This commit adds tasks in the 'debops.gitlab_runner' role which will
patch the 'vagrant-libvirt' source code to not remove SSH host keys
during packaging. This is hopefully a temporary solution and will be
fixed differently in the future.

Ref: https://github.com/vagrant-libvirt/vagrant-libvirt/issues/759