Redesign Python packaging for 'debops-tools'
This is a redesign of the Python 'lib/debops-tools/' package to be the
main 'debops' Python package again. This is done to synchronize the
"debops-tools" version and DebOps monorepo version. It should also make
installing 'debops' a bit more intuitive and accessible.

Source package creation using Python setuptools resulted in the Python
source files being put in the 'lib/debops-tools/debops/' subdirectory
instead of the 'debops/' subdirectory of the generated tarball,
therefore I moved the scripts to the root fo the monorepo. Otherwise the
current scripts were not changed; this shouldn't impact usage of the
Ansible roles/playbooks.

The source package generated by Python setuptools will now included
a LICENSE file, which was missing from the previous ones. The 'debops'
package description and classification was also updated. Package
versioning is done dynamically based on the 'git describe' result.