Include DebOps roles and playbooks in PyPI package
This should make the Python DebOps package on PyPI a bit more useful,
and provide simple, stable releases in the future. The DebOps monorepo
can still be installed manually if needed.

The DebOps scripts don't support the included roles and playbooks yet,
this will be implemented before the next release.