Fix Python package version information
Tests with installing generated Python source and wheel packages
revealed that the DebOps version is not correctly preserved in the
packages. These changes ensure that the version information retrieved
from the 'git describe' command is stored in the generated packages and
correctly re-used later.

Additionally, the custom 'debops/ansible' symlink should be correctly
created and removed as needed, fixing the issues during Python package
installation.