Check Python package builds on Travis-CI
The 'sdist' and 'wheel' Python packages were not built on Travis CI and
this caused a bug with the 'setup.py' script which worked in a Py2
environment, but not Py3.

Now, Travis CI will build the Python packages for testing purposes. They
are built in a quiet mode to avoid huge standard output, however this
should ensure that any issues with 'setup.py' will be caught early next
time.