[docs] Generate custom "Edit on GitHub" links
Some of the documentation files are generated from source '.yml' files
(for example default variable sections), some others are located
elsewhere in the monorepo (Changelog, installation documentation).

This commit implements a custom Python function that generates correct
URLs for the above set of files, used in the "Edit on GitHub" links in
the documentation.