Support roles/playbooks in Python package data dir
This patch implements support for using DebOps roles and playbooks
included in the Python package. The 'debops' scripts will check the path
for the 'debops' module installation directory and will use the Ansible
roles and playbooks found there, if the DebOps monorepo is not
installed.

This should make the installation of the DebOps monorepo optional, as
well as allow easy implementation of the stable DebOps releases.