Don't check Mailman configuration against PEP8
It looks like the Mailman 'mm_cfg.py' template cannot be easily modified
to satisfy 'pycodestyle' demands; most of the values in it are templated
directly using Jinja, and attempts to wrap it using json.loads() result
in Mailman service not starting properly.

Because of that, for the time being the 'mm_cfg.py' template will be
ignored by 'make pep8' test. This might be changed later, when a better
solution is found.