[debops.lxc] Provide reference Jessie template
The 'shellcheck' script is used to validate shell scripts included in
the Debops project. The LXC template scripts included in the
'debops.lxc' role pass the validation test, however this makes comparing
changes between Debian and DebOps versions of the script impossible.

To make comparsions possible, the role includes the Debian versions of
the LXC template scripts, with 'shellcheck' warnings and errors fixed.
They can be used to easily see what additions are present in the DebOps
versions of the LXC template scripts; they are not used during
configuration.