[debops.root_account] Fix 'no TTY' error
The default '/root/.profile' configuration file contains the 'mesg n'
command which gets called on each 'root' login. When a shell is executed
without corresponding TTY, for example via 'rsync+ssh' connection, the
above command results in an error:

    mesg: ttyname failed: Inappropriate ioctl for device

This patch adds a check for an existing TTY before running the 'mesg n'
command which should fix the issue.